Disable special treatment for en_GB (bsc#1008192)
When I enabled this, I thought there would mostly be missing strings
which could be filled in from en_US but apparently that was not the case
-- we had old outdated strings instead. However, I am not sure if that
was not maybe indeed my fault -- at some point I did change something
in the en_GB translation.

In any case -- not worth the bother, it seems.