Update RN sync config with new, unique package names (boo#1182045)
Due to a temporary configuration mishap that resulted in multiple
packages being available for download from OBS, we needed unique
package names.

Our release notes sync script downloads the RN package from OBS
with calls like this:

curl -s -S -o [OUTPUT_FILENAME].rpm \
  https://build.opensuse.org/public/build/Documentation:Auto/openSUSE_Leap_15.2/x86_64/_repository/release-notes-openSUSE

That means, it asks OBS for a package and takes the first package that
OBS provides rather than e.g. going through a list to see which of
potentially multiple packages with the same basic name may fit best.

Once a mishap has happened where two OBS packages build two RPMs of the
same name, we also don't have a lot of control over what OBS presents
to curl as the package it should use anymore. The only thing you can do
is ask for a wipe of the download repository but that does not seem to
be immediate.